Refactorizzaciónpara colocar la lógina en los services

+++ src/main/java/com/ferdev/restful/api/services/ProductServiceImpl.java
@@ -4,6 +4,8 @@ import com.ferdev.restful.api.entities.Product;
 import com.ferdev.restful.api.exceptions.ProductNotFoundException;
 import com.ferdev.restful.api.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,32 +21,50 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> getProducts() {
+    public List<Product> getProducts(String order) {
+        if (order != null && order.equals("price")) {
+            return this.productRepository.findByOrderByPriceAsc();
+        }
+
         return this.productRepository.findAll();
     }
 
     @Override
-    public List<Product> getProductsOrderedByPrice() {
-        return this.productRepository.findByOrderByPriceAsc();
-    }
-
-    @Override
-    public Optional<Product> getProductById(int id) {
-        return this.productRepository.findById(id);
+    public Product getProductById(int id) {
+        Optional<Product> optProduct = this.productRepository.findById(id);
+        if (optProduct.isPresent()) {
+            return optProduct.get();
+        } else {
+            throw new ProductNotFoundException("There is no product with id " + id);
+        }
     }
 
     @Override
     public Product createProduct(Product product) {
+        product.setId(0);
         return this.productRepository.save(product);
     }
 
     @Override
-    public void updateProduct(Product product) {
-        this.productRepository.save(product);
+    public Product updateProduct(Product product, int id) {
+        if (this.productRepository.findById(id).isPresent()) {
+            product.setId(id);
+            this.productRepository.save(product);
+            return product;
+        } else {
+            throw new ProductNotFoundException("There is no product with id " + id);
+        }
     }
 
     @Override
-    public void deleteProduct(int id) {
-        this.productRepository.deleteById(id);
+    public Product deleteProduct(int id) {
+        Optional<Product> optProduct = this.productRepository.findById(id);
+
+        if (optProduct.isPresent()) {
+            this.productRepository.deleteById(id);
+            return optProduct.get();
+        } else {
+            throw new ProductNotFoundException("There is no product with id " + id);
+        }
     }
 }
